+++ src/CollectionAbstract.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace DigipolisGent\Value;
 
 use \ArrayIterator;
@@ -21,24 +23,24 @@ abstract class CollectionAbstract extends ValueAbstract implements CollectionInt
     /**
      * Compare two Collection objects and check if they can be considered equal.
      *
-     * @param \DigipolisGent\Value\ValueInterface|\DigipolisGent\Value\CollectionInterface
-     *   Collection to compare this collection with.
+     * @param \DigipolisGent\Value\ValueInterface $object
+     *   Value object to compare this collection with.
      *
      * @return bool
      *   Both have the same values.
      */
-    public function sameValueAs(ValueInterface $collection)
+    public function sameValueAs(ValueInterface $object): bool
     {
-        /* @var $collection \DigipolisGent\Value\CollectionInterface */
-        if (!$this->sameValueTypeAs($collection)) {
+        /** @var \DigipolisGent\Value\CollectionInterface $object */
+        if (!$this->sameValueTypeAs($object)) {
             return false;
         }
 
-        if (!$this->sameCollectionCount($collection)) {
+        if (!$this->sameCollectionCount($object)) {
             return false;
         }
 
-        if (!$this->sameCollectionValues($collection)) {
+        if (!$this->sameCollectionValues($object)) {
             return false;
         }
 
@@ -50,7 +52,7 @@ abstract class CollectionAbstract extends ValueAbstract implements CollectionInt
      *
      * @return \ArrayIterator
      */
-    public function getIterator()
+    public function getIterator(): ArrayIterator
     {
         return new ArrayIterator($this->values);
     }
@@ -58,13 +60,13 @@ abstract class CollectionAbstract extends ValueAbstract implements CollectionInt
     /**
      * Check if the collection has the same amount of items as this collection.
      *
-     * @param \DigipolisGent\Value\CollectionInterface
+     * @param \DigipolisGent\Value\CollectionInterface $collection
      *   Collection to compare this collection with.
      *
      * @return bool
      *   Has the same amount of items.
      */
-    protected function sameCollectionCount(CollectionInterface $collection)
+    protected function sameCollectionCount(CollectionInterface $collection): bool
     {
         $collectionIterator = $collection->getIterator();
         return $this->getIterator()->count() === $collectionIterator->count();
@@ -73,13 +75,13 @@ abstract class CollectionAbstract extends ValueAbstract implements CollectionInt
     /**
      * Check if the collection has the same values as this collection.
      *
-     * @param \DigipolisGent\Value\CollectionInterface
+     * @param \DigipolisGent\Value\CollectionInterface $collection
      *   Collection to compare this collection with.
      *
      * @return bool
      *   Has the same values.
      */
-    protected function sameCollectionValues(CollectionInterface $collection)
+    protected function sameCollectionValues(CollectionInterface $collection): bool
     {
         $collectionIterator = $collection->getIterator();
         foreach ($this as $index => $item) {