+++ src/CollectionInterface.php
@@ -1,6 +1,10 @@
 <?php
 
+declare(strict_types=1);
+
 namespace DigipolisGent\Value;
+
+use IteratorAggregate;
 
 /**
  * Interface for the base ValueCollection object.
@@ -9,7 +13,7 @@ namespace DigipolisGent\Value;
  *
  * @package DigipolisGent\Value
  */
-interface CollectionInterface extends \IteratorAggregate, ValueInterface
+interface CollectionInterface extends IteratorAggregate, ValueInterface
 {
 
 }