+++ tests/CollectionAbstractTest.php
@@ -1,8 +1,12 @@
 <?php
 
+declare(strict_types=1);
+
 namespace DigipolisGent\Test\Value;
 
+use ArrayIterator;
 use DigipolisGent\Value\CollectionAbstract;
+use DigipolisGent\Value\CollectionInterface;
 use DigipolisGent\Value\ValueInterface;
 use PHPUnit\Framework\TestCase;
 
@@ -16,28 +20,34 @@ class CollectionAbstractTest extends TestCase
 
     /**
      * Test if the iterator method returns an ArrayIterator.
+     *
+     * @test
      */
-    public function testGetIterator()
+    public function getIteratorMethodReturnsArrayIterator(): void
     {
         $collection = $this->getMockForAbstractClass(CollectionAbstract::class);
-        $this->assertInstanceOf(\ArrayIterator::class, $collection->getIterator());
+        $this->assertInstanceOf(ArrayIterator::class, $collection->getIterator());
     }
 
     /**
      * Not the same if they are not of the same type.
+     *
+     * @test
      */
-    public function testTwoCollectionsAreNotSameValueIfDifferentTypes()
+    public function twoCollectionsAreNotSameValueIfDifferentTypes(): void
     {
-        $collection1 = $this->createCollectionMock('Type1', new \ArrayIterator());
-        $collection2 = $this->createCollectionMock('Type2', new \ArrayIterator());
+        $collection1 = $this->createCollectionMock('Type1', new ArrayIterator());
+        $collection2 = $this->createCollectionMock('Type2', new ArrayIterator());
 
         $this->assertFalse($collection1->sameValueAs($collection2));
     }
 
     /**
      * Not the same if they have not the same amount of values.
+     *
+     * @test
      */
-    public function testTwoCollectionsAreNotTheSameValueIfTheyHaveNotTheSameAmountOfValues()
+    public function twoCollectionsAreNotTheSameValueIfTheyHaveNotTheSameAmountOfValues(): void
     {
         $iterator1 = $this->createMock(\ArrayIterator::class);
         $iterator1
@@ -59,8 +69,10 @@ class CollectionAbstractTest extends TestCase
 
     /**
      * Not the same if the collection item keys are not the same.
+     *
+     * @test
      */
-    public function testTwoCollectionsAreNotTheSameIfArrayKeysAreDifferent()
+    public function twoCollectionsAreNotTheSameIfArrayKeysAreDifferent(): void
     {
         $value1 = $this->createMock(ValueInterface::class);
         $iterator1 = new \ArrayIterator([0 => $value1]);
@@ -75,8 +87,10 @@ class CollectionAbstractTest extends TestCase
 
     /**
      * Not the same if the values on the same index are not the same.
+     *
+     * @test
      */
-    public function testTwoCollectionsAreNotTheSameIfHaveNotTheSameValues()
+    public function twoCollectionsAreNotTheSameIfHaveNotTheSameValues(): void
     {
         $value1 = $this->createMock(ValueInterface::class);
         $value1
@@ -95,8 +109,10 @@ class CollectionAbstractTest extends TestCase
 
     /**
      * The same if the values on the same index the same.
+     *
+     * @test
      */
-    public function testTwoCollectionsAreTheSameIfHaveTheSameValues()
+    public function twoCollectionsAreTheSameIfHaveTheSameValues(): void
     {
         $value1 = $this->createMock(ValueInterface::class);
         $value1
@@ -124,7 +140,7 @@ class CollectionAbstractTest extends TestCase
      * @return \DigipolisGent\Value\CollectionInterface|\PHPUnit\Framework\MockObject\MockObject
      *   The mocked CollectionAbstract.
      */
-    protected function createCollectionMock($name, \ArrayIterator $iterator)
+    protected function createCollectionMock($name, \ArrayIterator $iterator): CollectionInterface
     {
         $mock = $this
             ->getMockBuilder(CollectionAbstract::class)